+++ src/main/kotlin/org/snd/mediaserver/model/MediaServerBookMetadataUpdate.kt
@@ -7,7 +7,7 @@ data class MediaServerBookMetadataUpdate(
     val title: String? = null,
     val summary: String? = null,
     val number: String? = null,
-    val numberSort: Float? = null,
+    val numberSort: Double? = null,
     val releaseDate: LocalDate? = null,
     val authors: List<MediaServerAuthor>? = null,
     val tags: Collection<String>? = null,
